+++ src/realm/db.cpp
@@ -2710,6 +2710,19 @@ LnkSetPtr Transaction::import_copy_of(const LnkSetPtr& original)
     return std::make_unique<LnkSet>();
 }
 
+CollectionBasePtr Transaction::import_copy_of(const CollectionBasePtr& original)
+{
+    if (!original)
+        return nullptr;
+    if (Obj obj = import_copy_of(original->get_obj())) {
+        ColKey ck = original->get_col_key();
+        return obj.get_collection_ptr(ck);
+    }
+    // return some empty collection where size() == 0
+    // the type shouldn't matter
+    return std::make_unique<LnkLst>();
+}
+
 std::unique_ptr<Query> Transaction::import_copy_of(Query& query, PayloadPolicy policy)
 {
     return query.clone_for_handover(this, policy);