+++ test/test_query2.cpp
@@ -2682,7 +2682,75 @@ TEST(Query_LinkCounts)
     CHECK_EQUAL(k1, match);
 }
 
-TEST(Query_Link_Minimum)
+struct TestLinkList {
+    ColKey add_link_column(TableRef source, TableRef dest)
+    {
+        return source->add_column_list(*dest, "linklist");
+    }
+    void create_object_with_links(TableRef table, ColKey col, std::vector<ObjKey> links)
+    {
+        LnkLst ll = table->create_object().get_linklist(col);
+        for (auto link : links) {
+            ll.add(link);
+        }
+    }
+    void add_links_to(TableRef table, ColKey col, ObjKey obj, std::vector<ObjKey> links)
+    {
+        LnkLst ll = table->get_object(obj).get_linklist(col);
+        for (auto link : links) {
+            ll.add(link);
+        }
+    }
+};
+
+
+struct TestLinkSet {
+    ColKey add_link_column(TableRef source, TableRef dest)
+    {
+        return source->add_column_set(*dest, "linkset");
+    }
+    void create_object_with_links(TableRef table, ColKey col, std::vector<ObjKey> links)
+    {
+        LnkSet ls = table->create_object().get_linkset(col);
+        for (auto link : links) {
+            ls.insert(link);
+        }
+    }
+    void add_links_to(TableRef table, ColKey col, ObjKey obj, std::vector<ObjKey> links)
+    {
+        LnkSet ls = table->get_object(obj).get_linkset(col);
+        for (auto link : links) {
+            ls.insert(link);
+        }
+    }
+};
+
+struct TestDictionaryLinkValues {
+    ColKey add_link_column(TableRef source, TableRef dest)
+    {
+        return source->add_column_dictionary(*dest, "linkdictionary");
+    }
+    void create_object_with_links(TableRef table, ColKey col, std::vector<ObjKey> links)
+    {
+        Dictionary dict = table->create_object().get_dictionary(col);
+        for (auto link : links) {
+            std::string key = util::format("key_%1", keys_added++);
+            dict.insert(Mixed(StringData(key)), Mixed(link));
+        }
+    }
+    void add_links_to(TableRef table, ColKey col, ObjKey obj, std::vector<ObjKey> links)
+    {
+        Dictionary dict = table->get_object(obj).get_dictionary(col);
+        for (auto link : links) {
+            std::string key = util::format("key_%1", keys_added++);
+            dict.insert(Mixed(StringData(key)), Mixed(link));
+        }
+    }
+
+    size_t keys_added = 0;
+};
+
+TEST_TYPES(Query_Link_Container_Minimum, TestLinkList, TestLinkSet, TestDictionaryLinkValues)
 {
     Group group;
     TableRef table1 = group.add_table("table1");
@@ -2701,8 +2769,9 @@ TEST(Query_Link_Minimum)
     auto k2 = table1->create_object().set_all(123, 123.f, 123.).get_key();
     auto k3 = table1->create_object().get_key();
 
+    TEST_TYPE test_container;
     TableRef table2 = group.add_table("table2");
-    auto col_linklist = table2->add_column_list(*table1, "linklist");
+    ColKey col_linktest = test_container.add_link_column(table2, table1);
 
     // table2
     // 0: { }
@@ -2710,143 +2779,54 @@ TEST(Query_Link_Minimum)
     // 2: { 1, 2 }
     // 3: { 1, 2, 3 }
 
-    LnkLst ll;
-    table2->create_object();
-    ll = table2->create_object().get_linklist(col_linklist);
-    ll.add(k1);
-    ll = table2->create_object().get_linklist(col_linklist);
-    ll.add(k1);
-    ll.add(k2);
-    ll = table2->create_object().get_linklist(col_linklist);
-    ll.add(k1);
-    ll.add(k2);
-    ll.add(k3);
+    test_container.create_object_with_links(table2, col_linktest, {});
+    test_container.create_object_with_links(table2, col_linktest, {k1});
+    test_container.create_object_with_links(table2, col_linktest, {k1, k2});
+    test_container.create_object_with_links(table2, col_linktest, {k1, k2, k3});
 
     Query q;
     TableView tv;
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).min() == 123;
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).min() == 123;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k2, tv.get_key(0));
     CHECK_EQUAL(k3, tv.get_key(1));
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).min() == 456;
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).min() == 456;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).min() == null();
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).min() == null();
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k0, tv.get_key(0));
 
-    q = table2->column<Link>(col_linklist).column<Float>(col_float).min() == 123.0f;
+    q = table2->column<Link>(col_linktest).column<Float>(col_float).min() == 123.0f;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k2, tv.get_key(0));
     CHECK_EQUAL(k3, tv.get_key(1));
 
-    q = table2->column<Link>(col_linklist).column<Float>(col_float).min() == 456.0f;
+    q = table2->column<Link>(col_linktest).column<Float>(col_float).min() == 456.0f;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 
-    q = table2->column<Link>(col_linklist).column<Double>(col_double).min() == 123.0;
+    q = table2->column<Link>(col_linktest).column<Double>(col_double).min() == 123.0;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k2, tv.get_key(0));
     CHECK_EQUAL(k3, tv.get_key(1));
 
-    q = table2->column<Link>(col_linklist).column<Double>(col_double).min() == 456.0;
+    q = table2->column<Link>(col_linktest).column<Double>(col_double).min() == 456.0;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 }
 
-TEST(Query_Link_Minimum_Set)
-{
-    Group group;
-    TableRef table1 = group.add_table("table1");
-    auto col_int = table1->add_column(type_Int, "int", /* nullable */ true);
-    auto col_float = table1->add_column(type_Float, "float", /* nullable */ true);
-    auto col_double = table1->add_column(type_Double, "double", /* nullable */ true);
-
-    // table1
-    // 0: 789 789.0f 789.0
-    // 1: 456 456.0f 456.0
-    // 2: 123 123.0f 123.0
-    // 3: null null null
-
-    auto k0 = table1->create_object().set_all(789, 789.f, 789.).get_key();
-    auto k1 = table1->create_object().set_all(456, 456.f, 456.).get_key();
-    auto k2 = table1->create_object().set_all(123, 123.f, 123.).get_key();
-    auto k3 = table1->create_object().get_key();
-
-    TableRef table2 = group.add_table("table2");
-    auto col_linkset = table2->add_column_set(*table1, "linkset");
-
-    // table2
-    // 0: { }
-    // 1: { 1 }
-    // 2: { 1, 2 }
-    // 3: { 1, 2, 3 }
-
-    LnkSet ll;
-    table2->create_object();
-    ll = table2->create_object().get_linkset(col_linkset);
-    ll.insert(k1);
-    ll = table2->create_object().get_linkset(col_linkset);
-    ll.insert(k1);
-    ll.insert(k2);
-    ll = table2->create_object().get_linkset(col_linkset);
-    ll.insert(k1);
-    ll.insert(k2);
-    ll.insert(k3);
-
-    Query q;
-    TableView tv;
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).min() == 123;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k2, tv.get_key(0));
-    CHECK_EQUAL(k3, tv.get_key(1));
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).min() == 456;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).min() == null();
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k0, tv.get_key(0));
-
-    q = table2->column<Link>(col_linkset).column<Float>(col_float).min() == 123.0f;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k2, tv.get_key(0));
-    CHECK_EQUAL(k3, tv.get_key(1));
-
-    q = table2->column<Link>(col_linkset).column<Float>(col_float).min() == 456.0f;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-    q = table2->column<Link>(col_linkset).column<Double>(col_double).min() == 123.0;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k2, tv.get_key(0));
-    CHECK_EQUAL(k3, tv.get_key(1));
-
-    q = table2->column<Link>(col_linkset).column<Double>(col_double).min() == 456.0;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-}
-
-TEST(Query_Link_MaximumSumAverage)
+TEST_TYPES(Query_Link_MaximumSumAverage, TestLinkList, TestLinkSet, TestDictionaryLinkValues)
 {
     Group group;
     TableRef table1 = group.add_table("table1");
@@ -2867,10 +2847,11 @@ TEST(Query_Link_MaximumSumAverage)
     (++it)->set_all(456, 456.f, 456.);
     (++it)->set_all(789, 789.f, 789.);
 
+    TEST_TYPE test_container;
     TableRef table2 = group.add_table("table2");
     auto col_double = table2->add_column(type_Double, "double");
     auto col_link = table2->add_column(*table1, "link");
-    auto col_linklist = table2->add_column_list(*table1, "linklist");
+    ColKey col_linktest = test_container.add_link_column(table2, table1);
 
     // table2
     // 0: 456.0 ->0 { }
@@ -2878,73 +2859,68 @@ TEST(Query_Link_MaximumSumAverage)
     // 2: 456.0 ->2 { 1, 2 }
     // 3: 456.0 ->3 { 1, 2, 3 }
 
-    LnkLst ll;
     auto k0 = table2->create_object().set_all(456.0, keys[0]).get_key();
     auto k1 = table2->create_object().set_all(456.0, keys[1]).get_key();
     auto k2 = table2->create_object().set_all(456.0, keys[2]).get_key();
     auto k3 = table2->create_object().set_all(456.0, keys[3]).get_key();
-    ll = table2->get_object(k1).get_linklist(col_linklist);
-    ll.add(keys[1]);
-    ll = table2->get_object(k2).get_linklist(col_linklist);
-    ll.add(keys[1]);
-    ll.add(keys[2]);
-    ll = table2->get_object(k3).get_linklist(col_linklist);
-    ll.add(keys[1]);
-    ll.add(keys[2]);
-    ll.add(keys[3]);
+
+    test_container.add_links_to(table2, col_linktest, k0, {});
+    test_container.add_links_to(table2, col_linktest, k1, {keys[1]});
+    test_container.add_links_to(table2, col_linktest, k2, {keys[1], keys[2]});
+    test_container.add_links_to(table2, col_linktest, k3, {keys[1], keys[2], keys[3]});
 
     Query q;
     TableView tv;
 
     // Maximum.
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).max() == 789;
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).max() == 789;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k2, tv.get_key(0));
     CHECK_EQUAL(k3, tv.get_key(1));
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).max() == 456;
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).max() == 456;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).max() == null();
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).max() == null();
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k0, tv.get_key(0));
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).max() == table2->link(col_link).column<Int>(col_int);
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).max() == table2->link(col_link).column<Int>(col_int);
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k1, tv.get_key(0));
     CHECK_EQUAL(k2, tv.get_key(1));
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).max() == table2->column<Double>(col_double);
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).max() == table2->column<Double>(col_double);
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 
 
-    q = table2->column<Link>(col_linklist).column<Float>(col_flt).max() == 789.0f;
+    q = table2->column<Link>(col_linktest).column<Float>(col_flt).max() == 789.0f;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k2, tv.get_key(0));
     CHECK_EQUAL(k3, tv.get_key(1));
 
-    q = table2->column<Link>(col_linklist).column<Float>(col_flt).max() == 456.0f;
+    q = table2->column<Link>(col_linktest).column<Float>(col_flt).max() == 456.0f;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 
 
-    q = table2->column<Link>(col_linklist).column<Double>(col_dbl).max() == 789.0;
+    q = table2->column<Link>(col_linktest).column<Double>(col_dbl).max() == 789.0;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k2, tv.get_key(0));
     CHECK_EQUAL(k3, tv.get_key(1));
 
-    q = table2->column<Link>(col_linklist).column<Double>(col_dbl).max() == 456.0;
+    q = table2->column<Link>(col_linktest).column<Double>(col_dbl).max() == 456.0;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
@@ -2954,47 +2930,47 @@ TEST(Query_Link_MaximumSumAverage)
     // Floating point results below may be inexact for some combination of architectures, compilers, and compiler
     // flags.
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).sum() == 1245;
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).sum() == 1245;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k2, tv.get_key(0));
     CHECK_EQUAL(k3, tv.get_key(1));
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).sum() == 456;
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).sum() == 456;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).sum() == table2->link(col_link).column<Int>(col_int);
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).sum() == table2->link(col_link).column<Int>(col_int);
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).sum() == table2->column<Double>(col_double);
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).sum() == table2->column<Double>(col_double);
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 
 
-    q = table2->column<Link>(col_linklist).column<Float>(col_flt).sum() == 1245.0f;
+    q = table2->column<Link>(col_linktest).column<Float>(col_flt).sum() == 1245.0f;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k2, tv.get_key(0));
     CHECK_EQUAL(k3, tv.get_key(1));
 
-    q = table2->column<Link>(col_linklist).column<Float>(col_flt).sum() == 456.0f;
+    q = table2->column<Link>(col_linktest).column<Float>(col_flt).sum() == 456.0f;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 
 
-    q = table2->column<Link>(col_linklist).column<Double>(col_dbl).sum() == 1245.0;
+    q = table2->column<Link>(col_linktest).column<Double>(col_dbl).sum() == 1245.0;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k2, tv.get_key(0));
     CHECK_EQUAL(k3, tv.get_key(1));
 
-    q = table2->column<Link>(col_linklist).column<Double>(col_dbl).sum() == 456.0;
+    q = table2->column<Link>(col_linktest).column<Double>(col_dbl).sum() == 456.0;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
@@ -3004,269 +2980,59 @@ TEST(Query_Link_MaximumSumAverage)
     // Floating point results below may be inexact for some combination of architectures, compilers, and compiler
     // flags.
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).average() == 622.5;
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).average() == 622.5;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k2, tv.get_key(0));
     CHECK_EQUAL(k3, tv.get_key(1));
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).average() == 456;
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).average() == 456;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).average() == null();
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).average() == null();
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k0, tv.get_key(0));
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).average() <
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).average() <
         table2->link(col_link).column<Int>(col_int);
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k2, tv.get_key(0));
 
-    q = table2->column<Link>(col_linklist).column<Int>(col_int).average() == table2->column<Double>(col_double);
+    q = table2->column<Link>(col_linktest).column<Int>(col_int).average() == table2->column<Double>(col_double);
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 
 
-    q = table2->column<Link>(col_linklist).column<Float>(col_flt).average() == 622.5;
+    q = table2->column<Link>(col_linktest).column<Float>(col_flt).average() == 622.5;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k2, tv.get_key(0));
     CHECK_EQUAL(k3, tv.get_key(1));
 
-    q = table2->column<Link>(col_linklist).column<Float>(col_flt).average() == 456.0f;
+    q = table2->column<Link>(col_linktest).column<Float>(col_flt).average() == 456.0f;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 
 
-    q = table2->column<Link>(col_linklist).column<Double>(col_dbl).average() == 622.5;
+    q = table2->column<Link>(col_linktest).column<Double>(col_dbl).average() == 622.5;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k2, tv.get_key(0));
     CHECK_EQUAL(k3, tv.get_key(1));
 
-    q = table2->column<Link>(col_linklist).column<Double>(col_dbl).average() == 456.0;
+    q = table2->column<Link>(col_linktest).column<Double>(col_dbl).average() == 456.0;
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 1);
     CHECK_EQUAL(k1, tv.get_key(0));
 }
 
-TEST(Query_Link_MaximumSumAverage_Set)
-{
-    Group group;
-    TableRef table1 = group.add_table("table1");
-    auto col_int = table1->add_column(type_Int, "int", /* nullable */ true);
-    auto col_flt = table1->add_column(type_Float, "float", /* nullable */ true);
-    auto col_dbl = table1->add_column(type_Double, "double", /* nullable */ true);
-
-    // table1
-    // 0: 123 123.0f 123.0
-    // 1: 456 456.0f 456.0
-    // 2: 789 789.0f 789.0
-    // 3: null null null
-
-    ObjKeys keys({3, 5, 7, 9});
-    table1->create_objects(keys);
-    auto it = table1->begin();
-    it->set_all(123, 123.f, 123.);
-    (++it)->set_all(456, 456.f, 456.);
-    (++it)->set_all(789, 789.f, 789.);
-
-    TableRef table2 = group.add_table("table2");
-    auto col_double = table2->add_column(type_Double, "double");
-    auto col_link = table2->add_column(*table1, "link");
-    auto col_linkset = table2->add_column_set(*table1, "linkset");
-
-    // table2
-    // 0: 456.0 ->0 { }
-    // 1: 456.0 ->1 { 1 }
-    // 2: 456.0 ->2 { 1, 2 }
-    // 3: 456.0 ->3 { 1, 2, 3 }
-
-    LnkSet ll;
-    auto k0 = table2->create_object().set_all(456.0, keys[0]).get_key();
-    auto k1 = table2->create_object().set_all(456.0, keys[1]).get_key();
-    auto k2 = table2->create_object().set_all(456.0, keys[2]).get_key();
-    auto k3 = table2->create_object().set_all(456.0, keys[3]).get_key();
-    ll = table2->get_object(k1).get_linkset(col_linkset);
-    ll.insert(keys[1]);
-    ll = table2->get_object(k2).get_linkset(col_linkset);
-    ll.insert(keys[1]);
-    ll.insert(keys[2]);
-    ll = table2->get_object(k3).get_linkset(col_linkset);
-    ll.insert(keys[1]);
-    ll.insert(keys[2]);
-    ll.insert(keys[3]);
-
-    Query q;
-    TableView tv;
-
-    // Maximum.
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).max() == 789;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k2, tv.get_key(0));
-    CHECK_EQUAL(k3, tv.get_key(1));
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).max() == 456;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).max() == null();
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k0, tv.get_key(0));
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).max() == table2->link(col_link).column<Int>(col_int);
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k1, tv.get_key(0));
-    CHECK_EQUAL(k2, tv.get_key(1));
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).max() == table2->column<Double>(col_double);
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-
-    q = table2->column<Link>(col_linkset).column<Float>(col_flt).max() == 789.0f;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k2, tv.get_key(0));
-    CHECK_EQUAL(k3, tv.get_key(1));
-
-    q = table2->column<Link>(col_linkset).column<Float>(col_flt).max() == 456.0f;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-
-    q = table2->column<Link>(col_linkset).column<Double>(col_dbl).max() == 789.0;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k2, tv.get_key(0));
-    CHECK_EQUAL(k3, tv.get_key(1));
-
-    q = table2->column<Link>(col_linkset).column<Double>(col_dbl).max() == 456.0;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-
-    // Sum.
-    // Floating point results below may be inexact for some combination of architectures, compilers, and compiler
-    // flags.
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).sum() == 1245;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k2, tv.get_key(0));
-    CHECK_EQUAL(k3, tv.get_key(1));
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).sum() == 456;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).sum() == table2->link(col_link).column<Int>(col_int);
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).sum() == table2->column<Double>(col_double);
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-
-    q = table2->column<Link>(col_linkset).column<Float>(col_flt).sum() == 1245.0f;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k2, tv.get_key(0));
-    CHECK_EQUAL(k3, tv.get_key(1));
-
-    q = table2->column<Link>(col_linkset).column<Float>(col_flt).sum() == 456.0f;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-
-    q = table2->column<Link>(col_linkset).column<Double>(col_dbl).sum() == 1245.0;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k2, tv.get_key(0));
-    CHECK_EQUAL(k3, tv.get_key(1));
-
-    q = table2->column<Link>(col_linkset).column<Double>(col_dbl).sum() == 456.0;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-
-    // Average.
-    // Floating point results below may be inexact for some combination of architectures, compilers, and compiler
-    // flags.
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).average() == 622.5;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k2, tv.get_key(0));
-    CHECK_EQUAL(k3, tv.get_key(1));
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).average() == 456;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).average() == null();
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k0, tv.get_key(0));
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).average() <
-        table2->link(col_link).column<Int>(col_int);
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k2, tv.get_key(0));
-
-    q = table2->column<Link>(col_linkset).column<Int>(col_int).average() == table2->column<Double>(col_double);
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-
-    q = table2->column<Link>(col_linkset).column<Float>(col_flt).average() == 622.5;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k2, tv.get_key(0));
-    CHECK_EQUAL(k3, tv.get_key(1));
-
-    q = table2->column<Link>(col_linkset).column<Float>(col_flt).average() == 456.0f;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-
-
-    q = table2->column<Link>(col_linkset).column<Double>(col_dbl).average() == 622.5;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k2, tv.get_key(0));
-    CHECK_EQUAL(k3, tv.get_key(1));
-
-    q = table2->column<Link>(col_linkset).column<Double>(col_dbl).average() == 456.0;
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 1);
-    CHECK_EQUAL(k1, tv.get_key(0));
-}
-
-TEST(Query_OperatorsOverLink)
+TEST_TYPES(Query_OperatorsOverLink, TestLinkList, TestLinkSet, TestDictionaryLinkValues)
 {
     Group group;
     TableRef table1 = group.add_table("table1");
@@ -3282,9 +3048,10 @@ TEST(Query_OperatorsOverLink)
     table1->get_object(keys[0]).set_all(2, 2.0);
     table1->get_object(keys[1]).set_all(3, 3.0);
 
+    TEST_TYPE test_container;
     TableRef table2 = group.add_table("table2");
     auto col_int2 = table2->add_column(type_Int, "int");
-    auto col_linklist = table2->add_column_list(*table1, "linklist");
+    ColKey col_linktest = test_container.add_link_column(table2, table1);
 
     // table2
     // 0:  0 { }
@@ -3295,12 +3062,8 @@ TEST(Query_OperatorsOverLink)
     auto k1 = table2->create_object().set_all(4).get_key();
     auto k2 = table2->create_object().set_all(4).get_key();
 
-    auto ll = table2->get_object(k1).get_linklist(col_linklist);
-    ll.add(keys[0]);
-
-    ll = table2->get_object(k2).get_linklist(col_linklist);
-    ll.add(keys[1]);
-    ll.add(keys[0]);
+    test_container.add_links_to(table2, col_linktest, k1, {keys[0]});
+    test_container.add_links_to(table2, col_linktest, k2, {keys[1], keys[0]});
 
     Query q;
     TableView tv;
@@ -3309,7 +3072,7 @@ TEST(Query_OperatorsOverLink)
 
     // Rows 1 and 2 should match this query as 2 * 2 == 4.
     // Row 0 should not as the power subexpression will not produce any results.
-    q = power(table2->link(col_linklist).column<Int>(col_int)) == table2->column<Int>(col_int2);
+    q = power(table2->link(col_linktest).column<Int>(col_int)) == table2->column<Int>(col_int2);
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k1, tv.get_key(0));
@@ -3317,7 +3080,7 @@ TEST(Query_OperatorsOverLink)
 
     // Rows 1 and 2 should match this query as 2 * 2 == 4.
     // Row 0 should not as the power subexpression will not produce any results.
-    q = table2->column<Int>(col_int2) == power(table2->link(col_linklist).column<Int>(col_int));
+    q = table2->column<Int>(col_int2) == power(table2->link(col_linktest).column<Int>(col_int));
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k1, tv.get_key(0));
@@ -3326,7 +3089,7 @@ TEST(Query_OperatorsOverLink)
 
     // Rows 1 and 2 should match this query as 2.0 * 2.0 == 4.0.
     // Row 0 should not as the power subexpression will not produce any results.
-    q = power(table2->link(col_linklist).column<Double>(col_dbl)) == table2->column<Int>(col_int2);
+    q = power(table2->link(col_linktest).column<Double>(col_dbl)) == table2->column<Int>(col_int2);
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k1, tv.get_key(0));
@@ -3334,7 +3097,7 @@ TEST(Query_OperatorsOverLink)
 
     // Rows 1 and 2 should match this query as 2.0 * 2.0 == 4.0.
     // Row 0 should not as the power subexpression will not produce any results.
-    q = table2->column<Int>(col_int2) == power(table2->link(col_linklist).column<Double>(col_dbl));
+    q = table2->column<Int>(col_int2) == power(table2->link(col_linktest).column<Double>(col_dbl));
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k1, tv.get_key(0));
@@ -3345,7 +3108,7 @@ TEST(Query_OperatorsOverLink)
 
     // Rows 1 and 2 should match this query as 2 * 2 == 4.
     // Row 0 should not as the multiplication will not produce any results.
-    q = table2->link(col_linklist).column<Int>(col_int) * 2 == table2->column<Int>(col_int2);
+    q = table2->link(col_linktest).column<Int>(col_int) * 2 == table2->column<Int>(col_int2);
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k1, tv.get_key(0));
@@ -3353,7 +3116,7 @@ TEST(Query_OperatorsOverLink)
 
     // Rows 1 and 2 should match this query as 2 * 2 == 4.
     // Row 0 should not as the multiplication will not produce any results.
-    q = table2->column<Int>(col_int2) == 2 * table2->link(col_linklist).column<Int>(col_int);
+    q = table2->column<Int>(col_int2) == 2 * table2->link(col_linktest).column<Int>(col_int);
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k1, tv.get_key(0));
@@ -3361,7 +3124,7 @@ TEST(Query_OperatorsOverLink)
 
     // Rows 1 and 2 should match this query as 2.0 * 2.0 == 4.0.
     // Row 0 should not as the multiplication will not produce any results.
-    q = table2->link(col_linklist).column<Double>(col_dbl) * 2 == table2->column<Int>(col_int2);
+    q = table2->link(col_linktest).column<Double>(col_dbl) * 2 == table2->column<Int>(col_int2);
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k1, tv.get_key(0));
@@ -3369,117 +3132,7 @@ TEST(Query_OperatorsOverLink)
 
     // Rows 1 and 2 should match this query as 2.0 * 2.0 == 4.0.
     // Row 0 should not as the multiplication will not produce any results.
-    q = table2->column<Int>(col_int2) == 2 * table2->link(col_linklist).column<Double>(col_dbl);
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k1, tv.get_key(0));
-    CHECK_EQUAL(k2, tv.get_key(1));
-}
-
-TEST(Query_OperatorsOverLink_Set)
-{
-    Group group;
-    TableRef table1 = group.add_table("table1");
-    auto col_int = table1->add_column(type_Int, "int");
-    auto col_dbl = table1->add_column(type_Double, "double");
-
-    // table1
-    // 0: 2 2.0
-    // 1: 3 3.0
-
-    ObjKeys keys({5, 6});
-    table1->create_objects(keys);
-    table1->get_object(keys[0]).set_all(2, 2.0);
-    table1->get_object(keys[1]).set_all(3, 3.0);
-
-    TableRef table2 = group.add_table("table2");
-    auto col_int2 = table2->add_column(type_Int, "int");
-    auto col_linkset = table2->add_column_set(*table1, "linkset");
-
-    // table2
-    // 0:  0 { }
-    // 1:  4 { 0 }
-    // 2:  4 { 1, 0 }
-
-    table2->create_object();
-    auto k1 = table2->create_object().set_all(4).get_key();
-    auto k2 = table2->create_object().set_all(4).get_key();
-
-    auto ll = table2->get_object(k1).get_linkset(col_linkset);
-    ll.insert(keys[0]);
-
-    ll = table2->get_object(k2).get_linkset(col_linkset);
-    ll.insert(keys[1]);
-    ll.insert(keys[0]);
-
-    Query q;
-    TableView tv;
-
-    // Unary operators.
-
-    // Rows 1 and 2 should match this query as 2 * 2 == 4.
-    // Row 0 should not as the power subexpression will not produce any results.
-    q = power(table2->link(col_linkset).column<Int>(col_int)) == table2->column<Int>(col_int2);
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k1, tv.get_key(0));
-    CHECK_EQUAL(k2, tv.get_key(1));
-
-    // Rows 1 and 2 should match this query as 2 * 2 == 4.
-    // Row 0 should not as the power subexpression will not produce any results.
-    q = table2->column<Int>(col_int2) == power(table2->link(col_linkset).column<Int>(col_int));
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k1, tv.get_key(0));
-    CHECK_EQUAL(k2, tv.get_key(1));
-
-
-    // Rows 1 and 2 should match this query as 2.0 * 2.0 == 4.0.
-    // Row 0 should not as the power subexpression will not produce any results.
-    q = power(table2->link(col_linkset).column<Double>(col_dbl)) == table2->column<Int>(col_int2);
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k1, tv.get_key(0));
-    CHECK_EQUAL(k2, tv.get_key(1));
-
-    // Rows 1 and 2 should match this query as 2.0 * 2.0 == 4.0.
-    // Row 0 should not as the power subexpression will not produce any results.
-    q = table2->column<Int>(col_int2) == power(table2->link(col_linkset).column<Double>(col_dbl));
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k1, tv.get_key(0));
-    CHECK_EQUAL(k2, tv.get_key(1));
-
-
-    // Binary operators.
-
-    // Rows 1 and 2 should match this query as 2 * 2 == 4.
-    // Row 0 should not as the multiplication will not produce any results.
-    q = table2->link(col_linkset).column<Int>(col_int) * 2 == table2->column<Int>(col_int2);
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k1, tv.get_key(0));
-    CHECK_EQUAL(k2, tv.get_key(1));
-
-    // Rows 1 and 2 should match this query as 2 * 2 == 4.
-    // Row 0 should not as the multiplication will not produce any results.
-    q = table2->column<Int>(col_int2) == 2 * table2->link(col_linkset).column<Int>(col_int);
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k1, tv.get_key(0));
-    CHECK_EQUAL(k2, tv.get_key(1));
-
-    // Rows 1 and 2 should match this query as 2.0 * 2.0 == 4.0.
-    // Row 0 should not as the multiplication will not produce any results.
-    q = table2->link(col_linkset).column<Double>(col_dbl) * 2 == table2->column<Int>(col_int2);
-    tv = q.find_all();
-    CHECK_EQUAL(tv.size(), 2);
-    CHECK_EQUAL(k1, tv.get_key(0));
-    CHECK_EQUAL(k2, tv.get_key(1));
-
-    // Rows 1 and 2 should match this query as 2.0 * 2.0 == 4.0.
-    // Row 0 should not as the multiplication will not produce any results.
-    q = table2->column<Int>(col_int2) == 2 * table2->link(col_linkset).column<Double>(col_dbl);
+    q = table2->column<Int>(col_int2) == 2 * table2->link(col_linktest).column<Double>(col_dbl);
     tv = q.find_all();
     CHECK_EQUAL(tv.size(), 2);
     CHECK_EQUAL(k1, tv.get_key(0));